chore: move update-applications file
